Fixed counter_currency value that was being saved

+++ lib/abokifx_indexer/models/currency_pair.rb
@@ -3,8 +3,8 @@ module AbokiFXIndexer
 
     attribute :base_currency, String
     attribute :counter_currency, String
-    attribute :buy_rate, Float
-    attribute :sell_rate, Float
+    attribute :buy_rate, BigDecimal
+    attribute :sell_rate, BigDecimal
     attribute :post_date, Date
     attribute :fetched_at, DateTime, default: DateTime.now
